[test][power] Add a local power monitor capture script.
Add a local python script which can be used to capture data from a
Monsoon power monitor.  The script is compatible with the current
mock-power-monitor capture script, however this one will actually talk
to a locally connected Monsoon and collect data.

It will account for the three main causes of clock drift which were
showing up in captured trace data.

1) Samples silently dropped by the capture script.
2) Calibration samples silently stripped form the data stream.
3) Monsoons fundamentally missing their nominal capture rate by many
   hundreds of ppm.

Additionally, it provide some options for how to handle dropped and
calibration samples, and for capturing the monitor's aux channel.
These features are not currently being used by the test framework,
but they could be leveraged in the future to provide better ways of
doing ongoing clock synchronization, and different ways to visualize
dropped or calibration samples.

Note that this script is useful for individual developers who happen
to have a Monsoon at their desk and are doing power investigations.
Automated testing infrastructure will need additional work in order to
integrate this script (or something like it) with their system.

Bug: 327634476

Change-Id: Icdb873c17ed1c88a21dfe64bc424f8ccad274214
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1005932
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Commit-Queue: John Grossman <johngro@google.com>