[vkext] Use config to choose what tests to run
Use configs to determine what GPU to use and whether to test for several
of features. This will allow us to set the configuration OOT when this
test is used as a subpackage. This also allows us to recombine
vkext_protected and vkext_unprotected binaries, since we can disable a
flag to run only the unprotected tests from the binary.

Change-Id: If9513ff2431cd6c2b512863c2a8964be198792f1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1014198
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Fuchsia-Auto-Submit: John Bauman <jbauman@google.com>
Reviewed-by: Josh Gargus <jjosh@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>