[fx][ssh] Moving sshconfig generation to tools/devshell
This moves the generation of the sshconfig file used
by `fx shell` to be done at runtime vs. as part of the
build.

This is done to enable using ffx configuration to manage the
ssh key path both in and out of tree in order to avoid mismatched
keys or confusion on which key is used when. Since the configuration
can change outside the build, it makes more sense to check it at runtime
vs. build time.

More info: go/fuchsia-ssh-unification

Bug: b/329486457

Change-Id: Ide76c4393b4cac340b92fb107fb53caf288f1410
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1001693
Reviewed-by: Oliver Newman <olivernewman@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>