[kernel][sched] Update RunQueue tiebreaker policy
This change updates (and officially documents) RunQueue's tiebreaker
story. This is made to conform to what we have in place in the kernel
today.

The policy now: the next thread is the one that is active and has the
earliest finish time within its current period. In the event of a tie of
finish times, the one with the earlier start time is picked - and in
then in the event of a tie of start times, the thread with the lowest
address is expediently picked (which is a small bias that should not
persist across runs).

Bug: 328641440
Change-Id: I457319b62db856e2c920cb468825705a78fed114
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1015555
Reviewed-by: Corey Tabaka <eieio@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>