[ufs] Support UFS power mode
This CL uses the STRAT STOP UNIT SCSI command and DME HIBERNATE UIC
command to change the power mode of a UFS device.
Since the suspend/resume of the device driver framework is not complete
yet, we verify it with a unit test.

Test: fx test ufs-tests
Change-Id: Iffb77e4b9039886340d3865ecab76987eacd9fc6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/939572
Reviewed-by: Christopher Anderson <cja@google.com>
Reviewed-by: HanBin Yoon <hanbinyoon@google.com>
Commit-Queue: Jaeyoon Choi <j_yoon.choi@samsung.com>