[build][bazel] Add @fuchsia_build_info repository.
A new repository, built from a custom repository rule,
used to hold information specific to the current Fuchsia
build configuration. This is setup without depending on
Ninja-generated outputs.

- Generate the content of the build_args.bzl file.

  Also change references from
  @legacy_ninja_build_outputs//:build_args.bzl into
  @fuchsia_build_info//:build_args.bzl instead.

- Compute Fuchsia checkout timestamps and hashes in
  the same way as //build/info/BUILD.gn, and use
  LINT comments to ensure they are kept in sync.

Note that the @fuchsia_sdk_common repository, shared between
the Fuchsia platform build, and the Bazel SDK test suite,
is not used here (as the test suite should not depend on
the way the platform build is configured, only on the
final IDK / Bazel SDK generated by it).

Bug: 322881092
Bug: 42074238
Change-Id: I071679ee33f881495f4a90b3fee3d7b39bb53091
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1016254
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: David Turner <digit@google.com>