[fdf] Don't destroy node on bind fail
Previously the node would be removed if
a driver failed to bind to it.  This was
an incorrect behavior that was not caught
in DFv2.

The test for this is the bind-fail-test,
which will check for the node existing when
we add the controller Allowlist.

Change-Id: I775c0ba927caf78c63e918d68c363aea2ce6db15
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1015134
Commit-Queue: Garratt Gallagher <garratt@google.com>
Reviewed-by: Sarah Chan <spqchan@google.com>