[debugger] Notify when new threads are discovered.
When a new thread is discovered by PopulateCurrentThreads(), send the
notification to the client.

This mostly affects Linux because we don't get separate thread
notifications for the current process. On Fuchsia, there may be some
edge cases where this is encountered, but I don't know of them off the
top of my head.

Improve message on assertion for unexpected thread notifications.

Change-Id: Ief1678e92380404b76943c2c6f360d8519f83598
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1017212
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Jacob Rutherford <jruthe@google.com>