[component] delivery: on_readable support in the framework
When crafting the sender in make_program_outgoing_router, we read the
delivery field and tell the sender to wait for readable if delivery is
`on_readable`.

The waiting is performed in a scope that is shutdown when unresolved, in
line with other places where we clear escrowed data on unresolve to
prevent cross-version ABI concerns.

Adds an integration test of a component using `delivery: on_readable`,
where it should not receive requests unless the client writes to the
server_end.

There's an unforeseen interaction between on_readable and the
CapabilityRequested hook. Because the CapabilityRequested hook overrides
the capability obtained from the outgoing directory, we also lose any
laziness property. Properly handling that is split to the next CL for
ease of review, where we'll also add tests re CapabilityRequested.

Bug: b/319754472

Change-Id: I1d5050badebb0aa3a346a6742153e6c88c0147e7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1015132
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Reviewed-by: Gary Bressler <geb@google.com>