[netstack3] Use IpDeviceSendContext in IGMP/MLD
The IGMP and MLD implementations currently depend on SendFrameContext,
which allows sending any kind of frame from a device, and the
implementation of it for IGMP and MLD specifically was, in practice,
delegating its implementation to IpDeviceSendContext, which is
specifically for sending IP packets.

Instead, depend on IpDeviceSendContext explicitly; this indicates in the
type system that the packets originating from IGMP/MLD are IP packets,
which will allow integrating the EGRESS IP layer filtering hook in the
implementation of IpDeviceSendContext more easily.

Test: netstack3-core-test
Change-Id: I833228a5d3011725c3b9e9254c544a8272a0eb63
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1014196
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>