[ffx][repository] Auto-pick target if unset
This implements partial auto-selection of targets: If no default target
is explicitly set, the repo server will attempt to auto-detect the
target and, if there is only one, serve packages to this target. If
multiple targets are detected, the repo server does not guess. In this
case, the user is required to explicitly set the target.

Fixed: b/329202124
Change-Id: Iac164b8a3a78e50f399d23e7fceb5753fa76b48a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1017034
Commit-Queue: Thomas Zander <thomasalpinus@google.com>
Reviewed-by: Sen Jiang <senj@google.com>