[fdf][devfs] Adding Controller allowlist mechanism
Adds an allowlist for each function of the
fuchsia.device/Controller interface.

Also provides a passthrough for the compat shim controller
interface, so the compat shim can be phased out incrementally.

This CL only adds the mechanism for restricting Controller
uses, but does not enforce any restriction.  A follow-up CL
will add the specific restrictions.

Testing: This should have not change in functionality, so current
tests of the Controller interface should continue to work.

Bug: 330573782

Change-Id: Iedd4cbe4747d073d204c32eaff6089d3c57e3a1e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1015393
Commit-Queue: Garratt Gallagher <garratt@google.com>
Reviewed-by: Novin Changizi <novinc@google.com>