[trace_processing] Add all threads to model.processes
Adds a map to store the pid->tid mapping while importing a trace.
This is used to then populate the TraceModel's process.threads field so that
all threads, including threads that don't have events, are saved.
This is needed so that later we can we can map the thread to a name, and to
its process, when we print out CPU metrics.

Also sets a default thread name if there isn't one.

Test using `fx test --host`

Bug: b/309000365, b/331457527

Change-Id: I9beae84d79ac5f1275635958847ad6cc917ee92f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1001908
Reviewed-by: Mark Seaborn <mseaborn@google.com>
Commit-Queue: Mary Xia <maryxia@google.com>
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>