[testrunner] Return reconnect failures as connection failures.
For the SSHTester, we check in the beginning of a test if the SSH client
is still connected in case it got disconnected during the previous test.
However, if reconnecting fails, we'd just return the error like a normal
test error. This should get returned as a connection error so that
testrunner's logic will try to reconnect and if it fails, it will exit
early instead of continuing to run tests.

Change-Id: I03d59e4df69e06faa1fdf5ab16584dd63ce26efe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1010752
Reviewed-by: Danielle Kay <danikay@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>