[swd-e2e-tests][ssh] Update docs for ssh key location
This updates the instructions to run the e2e tests to
use `ffx config get ssh.priv` to get the path of the private
ssh key.

This is part of the effort to remove the build-time ssh key
in favor of the runtime ssh key managed via `ffx config`.

For more information see go/fuchsia-ssh-unification

Change-Id: I3cec425dfc2acfb7f2874b242b10d9f567739110
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1016848
Reviewed-by: Ben Keller <galbanum@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>