[bt][rfcomm] Correctly set C/R bit in the Address Field
The Address Field of an RFCOMM packet contains a Command (C) / Response
(R) bit that signifies the type of packet being sent. The way it is set
depends on 1) The role of the device 2) Type of frame and 3) State of
the multiplexer.

Currently, the RFCOMM library does not use the frame type to make this
classification and instead uses the logic in GSM 5.2.1.2 to calculate
the C/R bit for all frame types. This is incorrect.

Update the lib to correctly set the C/R bit specifically when the type
of packet is a multiplexer command (UIH) by applying the classification
logic defined in GSM 5.4.3.1, which is a subclause of GSM 5.2.1.2.

Bug: b/330690839
Test: fx test bt-rfcomm-tests
Test: fx test bt-rfcomm-lib-tests

Change-Id: I61d9c846da420d53504fcc51eff5b2bedb79f37f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1009194
Reviewed-by: Marie Janssen <jamuraa@google.com>
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>