[driver_runtime] Update inlining stats for missed channel writes.
If a channel receives a message, but there is not a channel wait
registered, inlining will not happen. Later on when the channel
wait is registered, either it would show it was inlined,
or the wrong reason would be displayed.

This records the message as not inlined, then we make sure the
correct reason is reflected later in the stats.

Bug: 330742756
Change-Id: I0ff2a8ddcea274bd23d546be2d799c8f50032514
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1017446
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>
Reviewed-by: Novin Changizi <novinc@google.com>