[starnix] Fix ownership of proc files
Previously, /proc/pid/* files were given the fs uid of the owning
task.

The proc man page says that they should be owned by the effective
uid of the owning task.

In practice, Linux seems to give some of the directories under
/proc/pid/ effective uid ownership, and the rest of the files under
/proc/pid/ real uid ownership.  In lieu of an explanation for that
behavior, we're going with the euid for the moment, as it is the
intent expressed in the man page, and closer to what Linux actually
does.

This CL also adds a (disabled) test that ensures that the files in
/proc/pid/ retain the same inode number across invocations. This
passes on Linux.

Multiply: verity-benchmarks: 100
Change-Id: I3303d518f6255054406d436bf484685716905efd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1014460
Reviewed-by: Marco Vanotti <mvanotti@google.com>
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Jeremy Manson <jeremymanson@google.com>