[detect][tests] Add fake time to stop test flake
The integration test includes a test of the at-most-every logic. But
a test using the real clock flakes if there's a delay that causes the
condition to be checked after the backoff period expires.

This CL uses fake-time for that test.

Multiply: detect_integration_test
Fixed: b/325510737
Change-Id: Id5e54ddd5861db0196bb069dc61a496cd5eede95
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1009272
Commit-Queue: Chris Phoenix <cphoenix@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>