[Netcfg] Fix DHCPv4 Server "start_serving" bug
When Netcfg learns that a WlanAp interface has come online, it needs to
call `start_serving` on the interface's DHCPv4 server. This was done
correctly when the interface was observed to have gone from "offline" to
"online", but incorrectly when the interface is already online when
first discovered.

Fixed: b/331036445
Change-Id: I67d42c7490ed501998d214cc6dab97b8dfad2ef0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1018296
Commit-Queue: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Evan Wildenhain <wildenhain@google.com>
Reviewed-by: Antonio Linhart <antoniolinhart@google.com>