[sdk][vfs] Move Node class to `vfs` namespace
The `vfs::internal::Node` type is already a part of the public SDK VFS
API, and is used both in and out of tree. It is most prominent in the
`PseudoDir` type, which uses this base class for entry addition,
removal, and lookup. Therefore, it is moved into the top-level `vfs`
namespace.

Update enumeration constants to compliy with style guide.

Update all in-tree usages to avoid use of deprecated type aliases
and constants. Mark them as removed in the next API version to simplify
removal.

Test: No functional change.
Bug: b/311176363

Change-Id: Ia9c2a68ee4cb6fc3fdb5c4e9166666381c379eff
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/987053
Reviewed-by: Chris Suter <csuter@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>
API-Review: Chris Suter <csuter@google.com>
Reviewed-by: Yilong Li <liyl@google.com>
Reviewed-by: Gary Bressler <geb@google.com>