[component_manager] Fix start ordering bug that contributed to starnix flake
There is a race in start ordering where StartAction performs the
following sequence:

- Create a pending Runtime object
- Call ComponentRunner/Start with this runtime object
- Install the Runtime object in the component's execution state

The race is what happens if the component exits between (2) and (3). The
component exiting will trigger a StopAction. If the StopAction runs
before (3), it will effectively no-op since it didn't find the
component's Runtime. But this isn't correct because the component _did_
start. At the very least, it should dispatch a Stopped event.

To fix this bug, change the ordering to:

- Lock the ExecutionState
- Call ComponentRunner/Start with this runtime object
- Install the Runtime object in the component's execution state
- Unlock the ExecutionState

Because the runner is called under lock, the race can't happen because
StopAction will be forced to wait on the lock, and when it gets the lock
it will find the Runtime there.

This doesn't completely solve the flake, but hopefully mitigates it.

Bug: 325143035
Change-Id: Ib47dee7c3652a0022d9cdca6613c34c44490538d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1018357
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Gary Bressler <geb@google.com>