[dl] Introduce Module Unmapping
This CL unmaps the module's data when the Module data structure is
destroyed. This is usually performed when the module is unloaded by
`dlclose` but can also happen as part of the garbage collection process
during loading or at the conclusion of tests when the runtime dynamic
linker is torn down.

The implementation for dl::Unmap lives in separate TUs that get compiled
depending on the platform.

Bug: 331468188
Change-Id: I1666f6686e757af8879ed5551dee7e08262c5845
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1012878
Commit-Queue: Caslyn Tonelli <caslyn@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>