[ld] Consolidate build ID handling
The ld::ObserveBuildIdNote function existed but wasn't being used
or tested; it was a buggy implementation that was needlessly
complex to reuse elfldltl::ObserveBuildIdNote for little benefit.
So rewrite it to be simpler, and use it in all of bootstrap.h,
StartupLoad, and RemoteDecodedModule.  This can later be reused
by libdl too.

Bug: 42085436
Change-Id: I9a0c31d64a31c7c0c0da9d0bcaf1ecde472d4897
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1020512
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Caslyn Tonelli <caslyn@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>