[netstack3] Support routine control flow
Support the Jump and Return actions in fuchsia.net.filter that allow
filtering routines to jump to uninstalled routines or return to the
calling routine.

This change also clarifies the difference between Accept and Return. All
routines evaluate to a final Return action by default, meaning that they
allow execution to continue in the calling routine (or to other routines
in the same hook). The default verdict for an entire hook is Accept.
When an Accept action is triggered within a routine, this action is
propagated all the way to the installed chain from which it was reached,
and execution continues at the next chain installed in the same hook (if
any).

Fixed: 318718273
Test: netstack3-filter-test
Change-Id: Ia288dae258ad04399bf80e036f165d5484581bcf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1016954
Commit-Queue: Peter Johnston <peterjohnston@google.com>
API-Review: Bruno Dal Bo <brunodalbo@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Peter DeLong <pdelong@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>