[http-client] Add a no-op idle timeout configuration
In the next CL we would be teaching http-client to stop itself if it
idles for some time. It should be well tested but to prevent any
unintended breakage we'll use a structured config entry to specify if
and how long the timeout is, and only flip the flag selectively.

This CL adds a config and verifies that it correctly falls back to -1.
This works thanks to the fact that adding a component to a package also
adds all of its dependencies, so the config value is propagated to any
tests that use this component.

Change-Id: I34f2d9989e41dbf4e00161ccd579038902b591ee
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1019278
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Devon H. O'Dell <dhobsd@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>